Removed default test coverage, Extracted the overlay code into its own class.
